core: stdcm: add timetable manager for local timetables
