+++ core/src/main/java/fr/sncf/osrd/api/TimetableCacheManager.kt
@@ -3,43 +3,25 @@ package fr.sncf.osrd.api
 import com.google.common.collect.ImmutableRangeSet
 import com.google.common.collect.Range
 import com.google.common.collect.RangeSet
-import com.google.common.collect.TreeRangeSet
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonAdapter
-import com.squareup.moshi.Moshi
-import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
-import fr.sncf.osrd.api.conflicts.TrainRequirementsById
-import fr.sncf.osrd.conflicts.SpacingRequirement
 import fr.sncf.osrd.sim_infra.api.RawInfra
 import fr.sncf.osrd.sim_infra.api.ZoneId
-import fr.sncf.osrd.utils.json.UnitAdapterFactory
 import io.opentelemetry.api.trace.SpanKind
 import io.opentelemetry.instrumentation.annotations.WithSpan
 import java.nio.file.Files
-import java.time.Duration
-import java.time.Instant
-import java.time.ZonedDateTime
 import java.util.concurrent.ConcurrentHashMap
 import kotlin.io.path.Path
 import kotlin.io.path.exists
 import kotlin.io.path.readBytes
 import kotlin.io.path.writeBytes
-import kotlin.math.pow
 import kotlinx.coroutines.*
-import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.cbor.Cbor
-import okhttp3.OkHttpClient
-import okhttp3.Request
-import okhttp3.Response
 import org.slf4j.LoggerFactory
 
 typealias TimetableId = Int
-
-private const val PAGE_SIZE = 100
 
 @JvmInline
 value class STDCMRequirements(val map: Map<ZoneId, RangeSet<Double>>) :
@@ -85,11 +67,9 @@ value class STDCMRequirements(val map: Map<ZoneId, RangeSet<Double>>) :
  * EPOCH.
  */
 class TimetableCacheManager(
-    baseUrl: String,
-    authenticationHeader: String,
-    httpClient: OkHttpClient,
+    val timetableProvider: TimetableProvider,
     val localCacheLocation: String? = null,
-) : APIClient(baseUrl, authenticationHeader, httpClient) {
+) {
     private val cache = ConcurrentHashMap<TimetableId, STDCMRequirements>()
     private val mutexes = ConcurrentHashMap<TimetableId, Mutex>()
 
@@ -143,92 +123,13 @@ class TimetableCacheManager(
     ): STDCMRequirements {
         logger.info("Fetching timetable requirements for $timetableId")
 
-        val res = mutableMapOf<ZoneId, RangeSet<Double>>()
         val requirements =
             withLocalCache(localCacheLocation, "$timetableId.cbor") {
-                runBlocking {
-                    val requirements = fetchTrainRequirements(infraId, infra, timetableId)
-                    requirements.collect { spacingReq ->
-                        val set = res.computeIfAbsent(spacingReq.zone) { TreeRangeSet.create() }
-                        set.add(Range.closedOpen(spacingReq.beginTime, spacingReq.endTime))
-                    }
-                    STDCMRequirements(res)
-                }
+                timetableProvider.getTimetableRequirements(infraId, infra, timetableId)
             }
 
         logger.info("Saved timetable requirements for $timetableId")
         return requirements
-    }
-
-    @OptIn(ExperimentalCoroutinesApi::class)
-    private fun fetchTrainRequirements(
-        infraId: String,
-        infra: RawInfra,
-        timetableId: TimetableId,
-    ): Flow<SpacingRequirement> = flow {
-        val firstPageTrainRequirements = getTrainPaginatedRequirements(infraId, timetableId, 1)
-        emitAll(
-            firstPageTrainRequirements.results
-                .flatMap {
-                    it.spacingRequirements.map { spacingReq ->
-                        SpacingRequirement.fromRJSWithAddedTime(
-                            spacingReq,
-                            infra,
-                            it.startTime.durationSinceEpoch(),
-                        )
-                    }
-                }
-                .asFlow()
-        )
-        emitAll(
-            (2..firstPageTrainRequirements.pageCount)
-                .asFlow()
-                // Limit the number of concurrent calls to the requirements endpoint.
-                .flatMapMerge(concurrency = 5) { page ->
-                    flow {
-                        val paginatedTrainRequirements =
-                            getTrainPaginatedRequirements(infraId, timetableId, page)
-                        paginatedTrainRequirements.results
-                            .flatMap {
-                                it.spacingRequirements.map { spacingReq ->
-                                    SpacingRequirement.fromRJSWithAddedTime(
-                                        spacingReq,
-                                        infra,
-                                        it.startTime.durationSinceEpoch(),
-                                    )
-                                }
-                            }
-                            .forEach { emit(it) }
-                    }
-                }
-        )
-    }
-
-    private fun getTrainPaginatedRequirements(
-        infraId: String,
-        timetableId: TimetableId,
-        page: Int,
-    ): PaginatedRequirements {
-        val endpointPath = "timetable/$timetableId/requirements/"
-        val request =
-            buildRequest(endpointPath, "infra_id=$infraId&page=$page&page_size=$PAGE_SIZE")
-        val response = getWithRetries(request)
-        return paginatedRequirementsAdapter.fromJson(response.body.source())!!
-    }
-
-    /** Try to access a request, retries on error with increasing delay */
-    private fun getWithRetries(request: Request, nRetries: Int = N_RETRIES): Response {
-        var response: Response? = null
-        for (tryCount in 1..<nRetries) {
-            response = httpClient.newCall(request).execute()
-            if (response.isSuccessful) return response
-            else {
-                logger.error("Error when getting ${request.url}: $response")
-                val nextSleepDuration = 1_000 * 2.0.pow(tryCount).toLong()
-                Thread.sleep(nextSleepDuration)
-            }
-        }
-        throw UnexpectedHttpResponse(response)
     }
 
     /**
@@ -262,25 +163,4 @@ class TimetableCacheManager(
             return map
         }
     }
-
-    private data class PaginatedRequirements(
-        @Json(name = "page_count") val pageCount: Int,
-        val results: List<TrainRequirementsById>,
-    )
-
-    private val paginatedRequirementsAdapter: JsonAdapter<PaginatedRequirements> =
-        Moshi.Builder()
-            .addLast(UnitAdapterFactory())
-            .addLast(KotlinJsonAdapterFactory())
-            .build()
-            .adapter(PaginatedRequirements::class.java)
-}
-
-val EPOCH_ZONED: ZonedDateTime = Instant.EPOCH.atZone(java.time.ZoneId.of("UTC"))
-
-const val N_RETRIES = 5
-
-/** Returns the duration since EPOCH, in seconds, precise to the millisecond. */
-fun ZonedDateTime.durationSinceEpoch(): Double {
-    return Duration.between(EPOCH_ZONED, this).toMillis() / 1000.0
 }
